refactor: :recycle: extract RainbowService in new services package
